Adding finance calculator

+++ python_introduction/finance_calculator.py
@@ -1,11 +1,11 @@
 #Personal Finance Calculator
 
 #prompt the user to enter a monthly income and the total monthly expenses
-monthly_income = int(input("Enter your monthly income: "))
-total_monthly_expenses = int(input("Enter your total monthly expenses: "))
+monthly_income = float(input("Enter your monthly income: "))
+total_monthly_expenses = float(input("Enter your total monthly expenses: "))
 
 #calculate monthly savings
-monthly_savings = monthly_income - total_monthly_expenses
+monthly_savings = int(monthly_income - total_monthly_expenses)
 
 #convert monthly savings to whole number (remove decimal)
 #monthly_savings = int(monthly_savings)
